+++ changelog.md
@@ -4,9 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.3.2...HEAD)
+### Fixed
+* [PCM-1961](https://inindca.atlassian.net/browse/PCM-1961) ended softphone conversations are now immediately removed from the `current` array on the ISdkConversationUpdateEvent.
 
 # [v7.3.2](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.3.1...v7.3.2)
+### Added
 * [PCM-1856](https://inindca.atlassian.net/browse/PCM-1856) touched up and improved function for handling resolution changes
+
+### Fixed
 * [PCM-1948](https://inindca.atlassian.net/browse/PCM-1948) moved rxjs to be a full dependency rather than a dev dep.
 
 # [v7.3.1](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v7.3.0...v7.3.1)